backend/aiLingo/requirements.txt
django==4.2.11
djangorestframework==3.15.1
google-generativeai
requests
python-dotenv
djangorestframework-simplejwt
Pillow
django-cors-headers
gunicorn
psycopg2-binary
dj_database_url